# Conflicts:
#	tunnel/src/main/java/com/syscxp/tunnel/tunnel/TunnelManagerImpl.java

+++ billing/src/main/java/com/syscxp/billing/header/renew/RenewJob.java
@@ -1,6 +1,8 @@
 package com.syscxp.billing.header.renew;
 
 import com.syscxp.billing.header.sla.ProductCaller;
+import com.syscxp.core.CoreGlobalProperty;
+import com.syscxp.core.Platform;
 import com.syscxp.core.identity.InnerMessageHelper;
 import com.syscxp.core.rest.RESTApiDecoder;
 import com.syscxp.core.retry.Retry;
@@ -11,11 +13,8 @@ import com.syscxp.header.rest.RESTFacade;
 import com.syscxp.header.rest.RestAPIResponse;
 import com.syscxp.header.rest.RestAPIState;
 import com.syscxp.header.rest.TimeoutRestTemplate;
-import com.syscxp.header.tunnel.tunnel.APIUpdateExpireDateMsg;
-import com.syscxp.header.tunnel.tunnel.APIUpdateExpireDateReply;
+import com.syscxp.header.tunnel.tunnel.APIRenewAutoTunnelMsg;
 import com.syscxp.utils.gson.JSONObjectUtil;
-import org.quartz.JobExecutionContext;
-import org.quartz.JobExecutionException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.*;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -29,17 +28,13 @@ import com.syscxp.utils.logging.CLogger;
 import org.springframework.web.client.RestClientException;
 
 import java.io.IOException;
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.ListIterator;
-import java.util.Map;
 
 @Component
-public class RenewJob{
+public class RenewJob {
 
     @Autowired
     private DatabaseFacade dbf;
@@ -50,9 +45,11 @@ public class RenewJob{
     private TimeoutRestTemplate template;
 
     private static final CLogger logger = Utils.getLogger(RenewJob.class);
+
     public RenewJob() {
         template = RESTFacade.createRestTemplate(CoreGlobalProperty.REST_FACADE_READ_TIMEOUT, CoreGlobalProperty.REST_FACADE_CONNECT_TIMEOUT);
     }
+
     @Scheduled(cron = "0 0/5 * * * ? ")
     @Transactional
     protected void autoRenew() {
@@ -64,7 +61,7 @@ public class RenewJob{
             SimpleQuery<RenewVO> q = dbf.createQuery(RenewVO.class);
             q.add(RenewVO_.isRenewAuto, SimpleQuery.Op.EQ, true);
             List<RenewVO> renewVOs = q.list();
-            if(renewVOs == null){
+            if (renewVOs == null) {
                 logger.info("there is no activity renew product");
                 return;
             }
@@ -74,32 +71,35 @@ public class RenewJob{
                 RenewVO renewVO = ite.next();
                 LocalDateTime now = LocalDateTime.now();
                 LocalDateTime expiredTime = renewVO.getExpiredTime().toLocalDateTime();
-                if ( ChronoUnit.DAYS.between(now, expiredTime) > 7 ) {
+                if (expiredTime.isAfter(now)) {
+                    continue;
+                }
+                if ((ChronoUnit.DAYS.between(now, expiredTime) > 7) && expiredTime.isBefore(now)) {
                     dbf.getEntityManager().remove(dbf.getEntityManager().merge(renewVO));
                     dbf.getEntityManager().flush();
                     continue;
                 }
-                if(expiredTime.isAfter(now)){
-                    continue;
-                }
+
 
                 ProductCaller caller = new ProductCaller(renewVO.getProductType());
-                APIUpdateExpireDateMsg aMsg = caller.getCallMsg();
 
-                aMsg.setUuid(renewVO.getProductUuid());
-                aMsg.setDuration(1);
-                aMsg.setProductChargeModel(renewVO.getProductChargeModel());
-                aMsg.setType(OrderType.RENEW);
-                aMsg.setAccountUuid(renewVO.getAccountUuid());
-                InnerMessageHelper.setMD5(aMsg);
-                String gstr = RESTApiDecoder.dumpWithSession(aMsg);
-                RestAPIResponse rsp = syncJsonPost(caller.getProductUrl(), gstr,RestAPIResponse.class);
+                if (renewVO.getProductType().equals(ProductType.TUNNEL)) {
+                    APIRenewAutoTunnelMsg aMsg = new APIRenewAutoTunnelMsg();
 
-                if (rsp.getState().equals(RestAPIState.Done.toString())) {
-                    try {
-                        RESTApiDecoder.loads(rsp.getResult());
-                    } catch (Exception e) {
-                        logger.error(e.getMessage());
+                    aMsg.setUuid(renewVO.getProductUuid());
+                    aMsg.setDuration(1);
+                    aMsg.setProductChargeModel(renewVO.getProductChargeModel());
+                    aMsg.setAccountUuid(renewVO.getAccountUuid());
+                    InnerMessageHelper.setMD5(aMsg);
+                    String gstr = RESTApiDecoder.dump(aMsg);
+                    RestAPIResponse rsp = syncJsonPost(caller.getProductUrl(), gstr, RestAPIResponse.class);
+
+                    if (rsp.getState().equals(RestAPIState.Done.toString())) {
+                        try {
+                            RESTApiDecoder.loads(rsp.getResult());
+                        } catch (Exception e) {
+                            logger.error(e.getMessage());
+                        }
                     }
                 }
 
@@ -111,7 +111,8 @@ public class RenewJob{
 
     }
 
-    public <T> T syncJsonPost(String url, String body,  Class<T> returnClass) {
+
+    public <T> T syncJsonPost(String url, String body, Class<T> returnClass) {
         body = body == null ? "" : body;
 
         HttpHeaders requestHeaders = new HttpHeaders();
